Create two packages impl and move impl classes to those packages
